requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
asyncpg==0.31.0
alembic==1.18.4
passlib[argon2]==1.7.4
python-jose[cryptography]==3.5.0
pydantic-settings==2.13.1
python-multipart==0.0.20
boto3==1.40.19
aiosqlite==0.21.0